requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
black==21.7b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
discord.py==1.7.3
distlib==0.3.2
flake8==3.9.2
idna==3.2
isort==5.9.3
libxml2-python==2.9.12
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
pbr==5.6.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
regex==2021.8.3
requests==2.26.0
stevedore==3.3.0
tomli==1.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
